+++ common/src/main/java/net/mehvahdjukaar/polytone/Polytone.java
@@ -19,6 +19,7 @@ import net.mehvahdjukaar.polytone.tabs.CreativeTabsModifiersManager;
 import net.mehvahdjukaar.polytone.texture.VariantTextureManager;
 import net.mehvahdjukaar.polytone.utils.BiomeKeysCache;
 import net.mehvahdjukaar.polytone.utils.CompoundReloader;
+import net.mehvahdjukaar.polytone.utils.LazyHolderSet;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.toasts.SystemToast;
 import net.minecraft.client.gui.components.toasts.ToastComponent;
@@ -81,15 +82,13 @@ public class Polytone {
         sodiumOn = isSodiumOn;
         isDevEnv = devEnv;
         isForge = forge;
-        //TODO: rename effects, modifiers and properties to a common standard naming scheme
-        //TODO: colormap for particles
-        //for colormap grid support
 
-        //SKY and fog
-        //item properties and color. cache pllayer coord
+        //TODO: colormap for particles
+
+//TODO: colormap for paritcles? also register particlesproeprly, a registry removal stuff and also day time and weather chedk for emitters
+
         //exp color
         //biome lightmap
-        // per blockstate offset like block models
 
         if (isDevEnv) {// force all mixins to load in dev
             MixinEnvironment.getCurrentEnvironment().audit();
@@ -106,6 +105,7 @@ public class Polytone {
             BIOME_MODIFIERS.processAndApplyWithLevel(registryAccess, true);
             DIMENSION_MODIFIERS.doApply(registryAccess, true);
             BiomeKeysCache.clear();
+            LazyHolderSet.initializeAll(registryAccess);
         } catch (RuntimeException e) {
             Polytone.LOGGER.error("Failed to apply some Polytone modifiers on world load", e);