+++ neoforge/src/main/java/net/mehvahdjukaar/polytone/fluid/neoforge/FluidPropertiesManagerImpl.java
@@ -1,13 +1,11 @@
-package net.mehvahdjukaar.polytone.fluid.forge;
+package net.mehvahdjukaar.polytone.fluid.neoforge;
 
 import com.mojang.blaze3d.shaders.FogShape;
 import com.mojang.blaze3d.vertex.PoseStack;
-import net.mehvahdjukaar.polytone.Polytone;
 import net.mehvahdjukaar.polytone.fluid.FluidPropertyModifier;
 import net.mehvahdjukaar.polytone.utils.ColorUtils;
 import net.minecraft.client.Camera;
 import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.FogRenderer;
 import net.minecraft.core.BlockPos;