+++ neoforge/src/main/java/net/mehvahdjukaar/polytone/mixins/neoforge/CurioInventoryScreenMixin.java
@@ -1,16 +1,18 @@
-package net.mehvahdjukaar.polytone.mixins.forge;
+package net.mehvahdjukaar.polytone.mixins.neoforge;
 
 import net.mehvahdjukaar.polytone.slotify.SlotifyScreen;
-import net.minecraft.client.gui.screens.inventory.InventoryScreen;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyArg;
+import top.theillusivec4.curios.client.gui.CuriosScreen;
 
-@Mixin(InventoryScreen.class)
-public abstract class InventoryScreenMixin {
+@Pseudo
+@Mixin(CuriosScreen.class)
+public abstract class CurioInventoryScreenMixin {
 
     @ModifyArg(method = "renderBg", at = @At(value = "INVOKE",
-            target = "Lnet/minecraft/client/gui/screens/inventory/InventoryScreen;renderEntityInInventoryFollowsMouse(Lnet/minecraft/client/gui/GuiGraphics;IIIIIFFFLnet/minecraft/world/entity/LivingEntity;)V")
+            target = "Lnet/minecraft/client/gui/screens/inventory/InventoryScreen;renderEntityInInventoryFollowsMouse(Lnet/minecraft/client/gui/GuiGraphics;IIIFFLnet/minecraft/world/entity/LivingEntity;)V")
             , index = 1
     )
     public int modifyRenderEntityX(int x) {
@@ -25,7 +27,7 @@ public abstract class InventoryScreenMixin {
     }
 
     @ModifyArg(method = "renderBg", at = @At(value = "INVOKE",
-            target = "Lnet/minecraft/client/gui/screens/inventory/InventoryScreen;renderEntityInInventoryFollowsMouse(Lnet/minecraft/client/gui/GuiGraphics;IIIIIFFFLnet/minecraft/world/entity/LivingEntity;)V")
+            target = "Lnet/minecraft/client/gui/screens/inventory/InventoryScreen;renderEntityInInventoryFollowsMouse(Lnet/minecraft/client/gui/GuiGraphics;IIIFFLnet/minecraft/world/entity/LivingEntity;)V")
             , index = 2
     )
     public int modifyRenderEntityY(int y) {