+++ neoforge/src/main/java/net/mehvahdjukaar/polytone/mixins/neoforge/ForgeHooksClientMixin.java
@@ -1,0 +1,10 @@
+package net.mehvahdjukaar.polytone.mixins.neoforge;
+
+//@Mixin(ForgeHooksClient.class)
+public class ForgeHooksClientMixin {
+/*
+    @ModifyExpressionValue(method = "getFogColor", at = @At(value = "INVOKE", target = "Lnet/minecraftforge/client/extensions/common/IClientFluidTypeExtensions;modifyFogColor(Lnet/minecraft/client/Camera;FLnet/minecraft/client/multiplayer/ClientLevel;IFLorg/joml/Vector3f;)Lorg/joml/Vector3f;"))
+    private static Vector3f polytone$modifyWaterFogColor(Vector3f original, @Local FluidState fluidState) {
+        return FluidPropertiesManager.modifyFogColor(original, )
+    }*/
+}